delete old checkpoint code
