+++ torchtune/utils/_checkpointing/__init__.py
@@ -9,4 +9,4 @@ from ._checkpointer import (  # noqa
     FullModelMetaCheckpointer,
     FullModelTorchTuneCheckpointer,
 )
-from ._checkpointer_utils import ModelType  # noqa
+from ._checkpointer_utils import ModelType, transform_opt_state_dict  # noqa
